+++ renderComment.js
@@ -1,0 +1,27 @@
+import { eventeLikesButtons } from "./eventeLikeButtons";
+
+const listElement = document.getElementById("list");
+
+export const renderComment = ({ comments }) => {
+    const commentHtml = comments
+        .map((comments, index) => {
+            return `<li class="comment">
+    <div class="comment-header">
+      <div>${comments.name}</div>
+      <div>${comments.date}</div>
+    </div>
+    <div class="comment-body">
+      <div data-index="${index}" class="comment-text">${comments.text}</div>
+    </div>
+    <div class="comment-footer">
+      <div class="likes">
+        <span class="likes-counter">${comments.likes}</span>
+        <button data-index="${index}" class="like-button ${comments.active}"></button>
+      </div>
+    </div>
+  </li>`;
+        })
+        .join("");
+    listElement.innerHTML = commentHtml;
+    eventeLikesButtons();
+};
